+++ src/app/offer-form/offer-form.component.scss
@@ -1,21 +1,25 @@
-::ng-deep {
+:host {
+  display: block;
+
   .container {
     box-sizing: border-box;
   }
-  label {
-    display: block;
-    font-size: 0.8rem;
+
+  ::ng-deep {
+    label {
+      display: block;
+      font-size: 0.8rem;
+    }
+
+    mat-form-field {
+      width: 100%;
+    }
   }
 
-  mat-form-field {
-    width: 100%;
+  .publish {
+    display: flex;
+    align-items: flex-end;
+    justify-content: center;
+    padding-bottom: 22px;
   }
-
-}
-
-.publish {
-  display: flex;
-  align-items: end;
-  justify-content: center;
-  padding-bottom: 22px;
 }
