+++ src/app/offer-form/offer-form.module.ts
@@ -1,0 +1,20 @@
+import {NgModule,} from '@angular/core';
+import {MatFormFieldModule} from '@angular/material/form-field';
+import {MatIconModule} from '@angular/material/icon';
+import {FormsModule} from '@angular/forms';
+import {MatInputModule} from '@angular/material/input';
+import {MatButtonModule} from '@angular/material/button';
+import {OfferFormComponent} from "./offer-form.component";
+
+@NgModule({
+  declarations: [OfferFormComponent],
+  imports: [
+    FormsModule,
+    MatFormFieldModule,
+    MatIconModule,
+    MatInputModule,
+    MatButtonModule,
+  ],
+  exports: [OfferFormComponent],
+})
+export class OfferFormComponentModule {}
